Create video with the same name as bag file

+++ generatevideo.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
+if [[ $1 == "" ]]; then
+    echo Please provide bag file
+    exit 1
+fi
+
 TMPDIR=$(mktemp -d)
 
-rs-convert -i full_recording.bag -p $TMPDIR/ -c
-ffmpeg -r 30 -pattern_type glob -i "${TMPDIR}/*.png" video.mp4
+rs-convert -i $1 -p $TMPDIR/ -c
+ffmpeg -r 30 -pattern_type glob -i "${TMPDIR}/*.png" $(basename -s .bag "$1").mp4
 
 rm -rf $TMPDIR
